+++ zarrs_metadata/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add support for a `must_understand` field to `MetadataV3` (ZEP0009)
+  - Extensions can now be parsed in more than just the additional fields of array/group metadata (e.g. codecs)
+
 ## [0.3.5] - 2025-02-18
 
 ### Fixed